update perl bindings to use large fd sets
I've been sitting on this patch since late 2012; was digging through
some old systems and realized I'd never upstreamed this.  As I'm
getting ready to rebuild that system again, I figured... better late
than never?

This patch updates the perl bindings - namely select_info and
snmp_get_select_info.

The private binding, snmp_get_select_info(), now takes numfds as a
parameter.  If it receives a 0, it will size the
netsnmp_large_fd_set to the size of a normal fd set.  If it
receives a non-zero integer, it will allocate the appropriately
sized netsnmp_large_fd_set.

The public binding, SNMP::SNMP's select_info(), now takes an
optional parameter, $fdsize.  If not passed (existing code), it
defaults to zero, which is then passed to the private binding, so
the netsnmp_large_fd_set is sized like a normal fdset.  If it is
passed, however, the value is passed back to the private binding,
and you can use the large fdset.

This was/is needed by an application that makes use of ~5k
concurrent connections.

This patch was originally written on 2012-11-12 at 22:56 UTC.